pyproject.toml
[project]
name = "py-in-the-sky"
version = "0.1.0"
description = "Monte Carlo wealth and retirement simulation"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
